test: update update password form test
